refactor: move subclass into own file
