+++ src/controllers/content.cpp
@@ -6,11 +6,6 @@
 #include <string>
 #include <unordered_map>
 
-#include "models/character.hpp"
-#include "models/character_class.hpp"
-#include "models/character_race.hpp"
-#include "models/spell.hpp"
-
 void dnd::Content::printStatus() const noexcept {
     groups.printStatus();
     std::cout << "=== Spells ===\n";
