+++ src/controllers/content.hpp
@@ -8,6 +8,7 @@
 #include "models/character.hpp"
 #include "models/character_class.hpp"
 #include "models/character_race.hpp"
+#include "models/character_subclass.hpp"
 #include "models/spell.hpp"
 
 namespace dnd {
