+++ src/models/character_class.hpp
@@ -21,23 +21,12 @@ public:
     ) noexcept;
 };
 
-class CharacterSubclass : public FeatureHolder {
-public:
-    const std::string class_name;
-    CharacterSubclass(const std::string& name, std::vector<Feature>&& features, const std::string& class_name) noexcept;
-};
-
 inline CharacterClass::CharacterClass(
     const std::string& name, std::vector<Feature>&& features, const std::string& hit_dice,
     const std::vector<int>& asi_levels, int subclass_level
 ) noexcept
     : FeatureHolder(name, std::move(features)), hit_dice(hit_dice), asi_levels(asi_levels),
       subclass_level(subclass_level) {}
-
-inline CharacterSubclass::CharacterSubclass(
-    const std::string& name, std::vector<Feature>&& features, const std::string& class_name
-) noexcept
-    : FeatureHolder(name, std::move(features)), class_name(class_name) {}
 
 } // namespace dnd
