+++ src/parsing/models/character_subclass_file_parser.cpp
@@ -11,6 +11,7 @@
 #include <nlohmann/json.hpp>
 
 #include "models/character_class.hpp"
+#include "models/character_subclass.hpp"
 #include "parsing/models/feature_holder_file_parser.hpp"
 #include "parsing/models/spellcasting_feature_holder_file_parser.hpp"
 #include "parsing/parsing_exceptions.hpp"
